A simple first attempt at adding a feature to limit total spawning for a given asteroid set.
Use case: one could create a group corresponding to Bennu, and then use this feature to ensure there's only one Bennu at a time.  Plus: it's simple.  Minus: it only works if CA is allowed to rename asteroids, and it won't work if e.g. Bennu is renamed.  Better method: keep a table of asteroid GUIDs.